requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 설정 관리
pydantic-settings==2.1.0

# 데이터 처리
pandas==2.1.4
numpy==1.24.4
openpyxl==3.1.2

# AI 및 머신러닝
prophet==1.1.7
scikit-learn==1.3.2

# 시각화
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 웹소켓
websockets==12.0

# 캐시
redis==5.0.1
aioredis==2.0.1

# HTTP 클라이언트
httpx==0.25.2
aiohttp==3.9.1

# 유틸리티
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 로깅
structlog==23.2.0

# 환경 변수
python-dotenv==1.0.0
